score is now a double, add wikidata link

+++ VndbSharp/Models/VisualNovel/VisualNovel.cs
@@ -33,7 +33,7 @@ namespace VndbSharp.Models.VisualNovel
 		public ReadOnlyCollection<VisualNovelRelation> Relations { get; private set; }
 		public ReadOnlyCollection<TagMetadata> Tags { get; private set; }
 		public Single Popularity { get; private set; }
-		public UInt32 Rating { get; private set; }
+		public double Rating { get; private set; }
 		[JsonProperty("screens")]
 		public ReadOnlyCollection<ScreenshotMetadata> Screenshots { get; private set; }
 		public ReadOnlyCollection<StaffMetadata> Staff { get; private set; }
